Added additional fields to map to old CTF
Added the other fields necessary to map to the old ctf.  Not sure yet, but we may need to use the player model instead of just the user model to not confuse the game.  FOr our purposes I do not believe an admin mode is necessary at this time (though I could very well be proved wrong in the future...)